Refactor: Separate query keys and commands
Refactored the frontend to improve code organization and maintainability.

- Centralized TanStack Query keys into `lib/queryKeys.ts` to ensure consistency and reduce typos.
- Separated Tauri command wrappers from general utility functions by moving them from `lib/utils.ts` to a dedicated `lib/commands.ts` file.
- Updated all component files (`InteractionButton.tsx`, `ReleaseSelector.tsx`, and `hooks.ts`) to use the new query key constants and command wrappers.
- Updated `AGENTS.md` to reflect the new frontend file structure and guide future development.

a
