+++ cat-launcher/src/PlayPage/hooks.ts
@@ -1,7 +1,9 @@
-import { listenToReleasesUpdate, toastCL } from "@/lib/utils";
-import { updateReleasesForVariant } from "@/store/releasesSlice";
 import { useEffect } from "react";
 import { useDispatch } from "react-redux";
+
+import { listenToReleasesUpdate } from "@/lib/commands";
+import { toastCL } from "@/lib/utils";
+import { updateReleasesForVariant } from "@/store/releasesSlice";
 
 export function useReleaseEvents() {
   const dispatch = useDispatch();
